requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
lxml
python-dotenv
fastmcp
groq
pydantic

google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-docx

PyGithub


